Fix for #1666- Back button is not accessible via keyboard when navigating to the about page. (#1679)
* initial changes

* margin

* format

* resolve review comments

* Popup Content Height
